neighbour: use separate displacement module for each particle
In order to support multiple particle instances more correctly, the
neighbour module will need a separate displacement module for each
particle instance. For now, the GPU variant of from_binning does not
support two particle instances correctly, but all others do.

This also tests if the particle instances of the displacement modules
match with the provided particle instances in the Lua module of
the neighbour list implementation.