dynamics: add method connect() to TCF registration
The computation of the stress tensor autocorrelation requires
particle:aux_enable() to be called in sampling steps. The sampling
intervals, however, are unknown upon construction and not regular if
used in conjunction with blocking_scheme.

Let blocking_scheme:correlation() call tcf:connect() if such a method is
provided. It passes the sampling interval to the tcf module which
connects particle:aux_enable() to sampler.on_prepend_force. The
connection is managed (stored and disconnected) by blocking_scheme.

If sampler accepts arbitrary grids of sampling steps in the future, this
approach is easily extended: blocking_scheme constructs such a grid
and passes it to tcf:connect().