gpu: fix 'overcrowded placeholders' error
Using binning and neighbour list modules on the GPU could lead to the
error "overcrowded placeholders" followed by a simulation abort.

In order to be able to continue the simulation, resize the regarding
neighbour and binning arrays as soon as such an error is detected and
try to rebuild the lists with a lower occupancy in the same step.

Note that if the occupancy is too low, the simulation will abort
nevertheless.

This commit fixes #133206