mdsim::binning: fix memory access violoation in unit test
Changing the cell list occupancy also changes the number of binning
cells. Thus, the result of binning.ncell() is not the actual number of
cells finally used.

The memory access violoation happened because of a fixed number of
particle counters for each bin, which was too small in some cases.