neighbours: update list on reverse tag update of particle2
If the reverse tags (i.e. the ordering of the particle indices) of the
second particle instance changed, the neighbour lists where not updated,
although this is required as the mapping of array index<->particle index
becomes invalid.
Fix this by adding the reverse_tag_cache of particle2 to the neighbour
list cache.

Note that gpu/from_binning is not updated, as it does not support two
particle instances (yet).