excluded_volume: fix test, check code snippet in documentation
The test didn't do anything but defining a function (which was left
unevaluated).

Add a test for the code snippet quoted in the documentation. The code is
inlined from the test script.

For the second point, the test scripts need to be copied to the binary
tree and the documentation becomes dependent on the whole tree test/lua.
This is a disadvantage since modifications of a test script will trigger
the rebuild of the whole documentation. The best solution is probably to
add a CMake switch which allows one to not build the documentation.