make the thermodynamics unit test an integration test
The test suite unit/observables/thermodynamics did not only test the
functionality of the thermodynamics module, but was actually an
integration test validating the thermodynamic properties of a simple
Lennard-Jones fluid.

The test is refactored to call the simulation scripts
examples/liquid/lennard_jones.*.lua in a sequence. Afterwards, a small
C++ test case analyses the resulting H5MD file, performing the same
checks as before.

The parameters of the test correspond to the validation section in the
documentation. The tolerances are quite large, however, to avoid
excessively long run times of the tests.

Rescaling of the velocity after the equilibration run is missing. It may
not be crucial here, but could be included in the example scripts.