avoid exception in H5MD output of trajectories
If the sampling interval is omitted in the phase space writer, a dump of
the system state is triggered by the signals "start"and "finish" of
observables.sampler. These two signals are always emitted by
sampler:run(). Repeated calls lead to duplicate dumps at the _same_
simulation step, which is not valid H5MD.

The following workarounds are possible:

1) remove the signals "start", "finish" from sampler:run()

2) increment the step counter (but not the clock) after each call to
   sampler:run()

3) avoid connection of the phase space writer to start/finish by
   explicitly passing a sampling interval, e.g., the total number of
   simulation steps.

This patch applies the last option and fixes #286246.