cmake: use one list to switch potentials on or off
As the number of potentials grows, a single checkbox for each one will
clutter the CMake configuration file. Instead we use a single list which
is used to define non-cached variables HALMD_WITH_${potential} in the
top-level CMake script.

Drop the special treatment of the Lennard-Jones potential. Disable tests
which link against this potential. (Script-based tests will fail if
LJ is disabled.)

lua: Copy only these potential modules which are enabled

doc: FIXME The doc still resorts to the Lua files in the source tree and
     not in the build tree. (I suppose that the order of execution of
     the various CMake subdirectories is undefined.) Thus the
     documentation will still list disabled potentials as well.