integrators (gpu): reduce memory access, fix bug with g_image
For most particles, the image vector is likely to remain unchanged in
one invocation of the kernel. Thus, reading and writing the image vector
in any case wastes bandwidth.

The change revealed a severe bug in the Andersen and Nosé-Hoover
thermostats: the image vector was never written back. The bug is
probably only relevant if tagged particle displacements over long times
are considered, e.g., in the computation of the MSD. In this case, the
unmodified dynamics (i.e., without thermostat) should be used.

Nevertheless, it is embarrassing that the bug was not detected by any of
the unit tests.

Results from the Lennard-Jones benchmark (before/after the patch) for
the first step of the velocity-Verlet integrator (64,000 particles):

    Tesla T10:  255 µs → 215 µs
    Tesla K20:   95 µs →  85 µs