mdsim: add region module to select particles from a simulation domain
This module allows to select particles based on their position in the
simulation domain. Whether a particle is within a region or outside of
it, will be determined by another module, the geometry.

The GPU implementation of the selection uses the copy_if() function with
an appropriate predicate which loads the position and calls the geometry
functor. With this implementation only one region (particles "inside"
or "outside" the geometry) can be selected.

A particle mask can also determined independently in order to enable operations
like intersections, unions, etc. of regions/particle groups.

Note that there is not yet a geometry implementation, this will follow
in a later commit.