particle_group: add function to copy particle group into particle
This function will copy only position, image, velocity, mass, species.
Others do not make much sense in a new particle instance (e.g. id, force
or auxiliary variables).

In order to make this available for the group, a binding to
to_particle() in each particle_group is required. This can only happen
in the particle group itself, because knowledge about the particle
instance is required and this is not available in the base class.

Refs #198261