dynamics/blocking_scheme: call aux_enable() only if required
Instead of letting the tcf call particle:aux_enable(), let the
blocking scheme handle it, as it knows (through `next()') when
a sampling step occurs. If the correlation module has, aux_enable
defined, particle:aux_enable() will be called automatically then.

This is a follow-up for 02ed3a22a9c and reverts it.

Refs #282179