test/…/two_particles: disable binning on GPU
Due to the small system size, using the neighbour list with binning will
always fail (at least when using the default particle numbers). Therefore,
disable the binning for the GPU.

A more optimal solution would be a check in the neighbour module itself,
allowing the usage of from_particle() in case the binning and neighbour
module are incompatible.