neighbour list: fall back to from_particle, if binning is incompatible
If the binning module has less than 3 cells in any spatial direction,
the neighbour list cannot constructed on the GPU using binning. Provide
a function in the from_binning module in order to test compatibility.

If this compatibility test fails, the lua module will fall back to
from_particle automatically. This removes the need for the user to
disable binning for small systems manually (as done in
test/…/two_particles).

Closes #412998