halmd/script: search directory of simulation script for Lua packages
This allows one to use custom modules in simulation scripts, e.g., for
factoring out the definitions of interactions shared by thermalisation
and production scripts. So far, only the current working directory was
included in the Lua search path.

For this it was necessary to revise package_path() and package_cpath()
of halmd::script:

Put the (complicated) calls to the Lua library in the private member
functions. Let the caller (here, the constructor) decide which path is
prepended. This improves readability and re-usability.