thermodynamics test: rescale velocities after thermalisation
This change is supposed to fix the frequently broken test suite.

The internal energy at the end of the NVT thermalisation run typically
deviates from its long-time average. Rescale the particle velocities
such that the initial state of the equilibration run precisely matches
the average internal energy. Instead of "measuring" the average energy
from the output of the thermalisation run automatically, we provide the
value from manual simulations.

Note that a naïve velocity rescaling was implemented before in the
Boltzmann thermostat (by constraints on the random velocities), which
was removed during the development.