Workaround trajectory test error.
Gives the h5md files of the trajectory test unique filenames in order
to avoid the "unable to truncate a file which is already open" error
produced by hdf5.

This is a quick and dirty workaround, but the core issue should be solved
when switching to a current version of h5xx anyways.