phase_space: one-step reading of samples
Adds a new kind of on_read functions to halmd/io/readers/* that takes a setter
slot function. Those are used as a replacement for the old procedure for
setting the contents of a host phase_space sample. This will hopefully make
it easier to split the phase_space sample and move H5MD reading functionality
away from phase_space.