phase_space: initial split of phase space samples
Splits the phase space sample into new generic sample classes.
The original samples::phase_space class is retained for now,
but mainly acts as a wrapper around samples::sample classes.

The clock steps are still propagated, but are functionally
replaced by cache observers (yet to be done in the GPU
implementation).

The affected unit tests are provisionally updated, but still
partly use the deprecated samples::phase_space class and have
to be updated accordingly.

The type of H5MD data sets is exposed to lua, so that the
phase_space reader can create correctly typed instances of the
samples::sample classes while reading a file.

The phase_space reader utilizes the new sample classes, but
still reads to a samples::phase_space object and has to be
updated along with the reading procedure in general.