phase_space: remove obsolete samples::phase_space::set function
Removes the obsolete samples::phase_space::set function. It is
replaced by the individual samples::phase_space::set_* functions.

The phase_space reader now does not return a phase space sample,
but a table of individual samples equipped with a metatable to
access meta-information (i.e. nparticle, dimension, nspecies).

The lua implementation of phase_space still retains a set function
that now takes a table of samples (as obtained by the reader) and
calls samples::phase_space::set_* accordingly. This way the lua
interface for reading samples and writing them to a particle instance
can remain the same for now.