forces/potentials: move smoothing kernel into first potential adapter
Replaces mdsim/forces/* with a first simple potential adapter local_r4
that automatically applies the smoothing function. Removes all uses of
the old smoothing kernels, adjusts tests and examples and adds an
initial potential:truncate(...) interface to the lua interface of the
potentials. Cutoff remains in the potentials for now.