potential adapters: add hard_core modification
Adds a hard_core potential adapter. Moves the truncate interface to a
new adapters.lua module shared by all potentials and adds a modify interface.

The hard_core adapter is only instantiated for variants of the power_law
potential for now and the power_law_with_core potential is retained
as it contains an optimized version if the modified potential.

This commit refs #281801