lua: add main script
Adds a main script that is called internally by every invocation of halmd
and calls the "run" function that is now to be defined by every simulation
scripts (the examples and tests have been adjusted accordingly).
The script performs some initial setup and calls the optional define_args
function in the simulation script, if present.
The script is installed along with the other lua modules and can hence
be modified by the user to adjust the default invocation settings for
user scripts, if so desired.