test: add CUDA memory checks
Adds a macro halmd_add_gpu_test to be used for adding tests that
run on the GPU. This will automatically use cuda-memcheck to
validate the GPU memory usage during the test run. Optionally the
first argument to halmd_add_gpu_test can be NO_MEMCHECK to disable
the memory check for specific tests (the distinction between CPU
and GPU is still maintained to make it possible to add further
CUDA specific modifications to the tests in the future).

This commit changes add_test to halmd_add_gpu_test for all
existing GPU tests.

Furthermore a CMAKE configuration variable HALMD_TEST_CUDA_MEMCHECK
(defaults to ON) is defined, which can be used to disable
the memory checks for all tests.