doc: add documentation for the new simulation scripts
Adds a section "Usage/Simulation Scripts" in the documentation that
explains the basic structure of simulation scripts.
Add a notice to the documentation of the argument parser that refers
back to the Simulation Scripts section, as manual creation of
an argument parser is replaced by the "define_args" entry point.