cmake: remove -DNDEBUG from compile flags in RelWithDebInfo mode
This enables assert() and LOG_DEBUG/LOG_TRACE, which is otherwise only
available in Debug mode, i.e., building without any optimisation.

However, there is a debate whether assertions should be used at all:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=711515

On the other hand, it is our own decision which kind of log messages are
included if NDEBUG is defined.