dsfloat: initial dsfloat specialization for (gpu) particle
Adds specialized data types dsfloat_ptr and dsfloat_vector and
replaces the compile-time flag HALMD_VARIANT_VERLET_DSFUN with
explicit specializations of particle for float and dsfloat.
The precision is then decided at runtime in particle.lua.in
via an optional "precision" argument.
The implementation is only partial, as only particle itself and the
immediate dependencies are updated - the remainder of HALMD is still
missing explicit specialisations for dsfloat, so in the current state
only single precision will pass the tests.