fix: use potential truncation in the benchmark scripts
The benchmark scripts still used the old syntax for the potential truncation
resulting in untruncated potentials connected with a significant performance
loss. This commit modifies the scripts to use the new potential adapter
truncation.
Also modifies the potential adapter script to avoid a potential lua error
if no truncation type is given (the default is "shifted").