workaround for clang bug
Replaces the default initialization of std::function<void()> update_function
in particle_array_gpu by an empty function object and assign the member variable
to an empty lambda function in the constructor body if the function object
is empty.

This works around a bug that leads to a crash in clang.
A bug report has been sent to the clang mailing list.