particle: increase block size if otherwise grid size exceeds maximum
Increases the default block size in particle if the resulting grid size
would exceed the maximal grid size supported by the cuda device.

Since for complex kernels a lower block size is better than a larger block
size, the default block size in particle is kept as low as possible.

In the future a more fine-grained per-kernel estimation of an adequate
block size may result in performance gains.