h5md: writer: don't overwrite by default
The h5md writer does not overwrite by default anymore. Instead
it can optionally be requested to overwrite by an additional optional
flag in the C++ interface or by the optional argument overwrite in
the lua interface.

The lua interface of the observables has been adjusted to pass the overwrite
flag to the h5md file constructor and the equilibration tests are supplemented
by an additional "--overwrite" argument to allow overwriting their output
(this is needed for the thermodynamics tests and serves as an example use case
of the new behavior).