cuda: add fixed_total_threads argument to configure_kernel and respect min_grid_size
Updates configure_kernel to take an additional argument fixed_total_threads that allows
to enforce a fixed total number of threads (true) or to request a kernel configuration with
at least the specified number of threads (false).

Additionally if the grid size for the estimated optimal block size is lower than the
minimal grid size to achieve a good device occupancy the default configuration, which
uses a minimal block size, is used instead.