thermodynamics test: fix validation code
Fix reading of centre-of-mass velocity from H5MD file. Since the
dimension is a runtime variable, a fixed_vector<..., 3> was used, but
the 3rd component was not properly initialised to zero. Also fix the
flawed pointer recast (the address of container is not necessarily the
address of the data.)

The bug was spotted by occasionally failing nightly tests for 2 dimensions:

…/halmd/test/integration/validation/thermodynamics.cpp(116): error: in "validation": absolute value of norm_inf(v_cm){1.1185995965152789e+253} exceeds 3.5762786865234375e-06