density_mode: increase MAX_BLOCK_SIZE to 1024
To utilise the parallelity of newer GPUs (e.g. GeForce 1080), it is
tempting to increase DEVICE_SCALE to 4 to run 1024 threads per block in,
e.g., reduction algorithms. (Note that good performance also requires
that sufficiently many blocks are launched.) This change lead to the
following error in observables::gpu::density_mode, which is fixed here.
All other instances of a hard-coded block size include already 1024.

[11-03-2019 15:39:49.592384] [ERROR] density_mode (A): failed to compute density modes on GPU
[11-03-2019 15:39:49.722647] [ERROR] [CUDA] an illegal memory access was encountered