gpu::phase_space: fix severe bug in the cache observing logic
The host samplers were observing the state of phase_space_host_cache to decide
whether an update is needed or not. However, as different host samplers can
share the same host cache (e.g., for position, species), the cache appears
fresh to the second sampler. This lead to old data in the host samplers, e.g.,
for output to a file.

Solution: let the host samplers observe directly the underlying particle array on the GPU.