utility::wavevector: fix wrong output of wavenumbers
In some cases, the wavenumbers stored along with, e.g., the static
structure factor (observables.ssf) were wrong. This happened when some
of the requested wavenumbers were discarded because of incompatibility
with the reciprocal lattice. In the output to, e.g., H5MD files, some
values were deleted wrongly.

The reason was that erase() on a std::vector was walking from the first
to the last element, thereby invalidating all subsequent iterators. In
practice, these iterators were each pointing to their next element.
Besides the wrong output this had the potential for segfaulting.

The solution is to walk backwards over the list to be discarded.

This is a backport of commit:ae57ddf0e.