Make CMake use CUDA shared runtime library by default
When using CUDA 9.2 most of the GPU tests would fail with the error message:
Program hit cudaErrorInvalidDeviceFunction (error 8) due to "invalid device function" on CUDA API call to cudaLaunch.

Tracking down the issue we notices that it had something to do with the order of the files when linking the program.
We also notices that the bug would not appear when using the shared runtime library, which this post confirms:
https://stackoverflow.com/questions/57915085/cmake-cuda-invalid-device-function-even-with-correct-sm-version

To fix the error HALMD now uses the shared runtime library by default. To use static linking anyway
use the standard CMake flag CUDA_USE_STATIC_CUDA_RUNTIME and turn it ON.
The flag HALMD_USE_STATIC_LINKING does not change this behavior, which the script will inform you about.
There is also a warning when using CUDA_USE_STATIC_CUDA_RUNTIME ON together with CUDA 9.2.

Also changes the documentation cache_vars.rst.in and faq.rst.in accordingly.