Implement new cuda::device class
The halmd::device class now holds a static instance of cuda::device as this is
needed with the new cuda-wrapper update.

All calls to cuda::device::get were replaced with halmd::device::get becuase
cuda::device::get is not static anymore.

The new command line argument --gpu-device can be used to select a GPU, if it
is not given the first available GPU will be chosen.