Use new cuda::texture class in force kernels
Because of the complicated structure of the force kernels and the kernels
which are calling them implementing the new cuda::texture class for these
was comlicated.

To pass the cudaTextureObject_t into the kernel the potential kernel classes
are used. This was necessary because of the inheritance of some of the classes
and the way the kernels are called (from pair_*.hpp).

The texture is now stored inside the potential kernel classes which means
that an instance of these (with the texture set) has to be passed to the
kernel. Instead of calling the constructor inside the kernel the kernel now
calls potential::fetch. The potential::operator() can be called just like
before.
A potential instance with the proper texture already set is returned
by the potential_class::get_gpu_potential function. This instance can be passed
direcly into the kernel.