Consider shared memory size when calculating maximum block size
Before this commit the shared memory size could be exceeded when
calculating the maximum block size in `configure_kernel` because this was
just not considered.

Now the block size can't be chosen bigger than the maximum block size which
doesn't exceed the shared memory required.

Use a fixed shared memory size of 48 kB as a temporary solution.