examples: simplify LJ scripts
There is only a single species, so we don't need to handle the case of
several species.

In H5MD output and input files, don't include data for species and mass.

Still demonstrate the use of group selection by a geometry, but use a
sphere rather than a cuboid, which is nicer to set up and perhaps more
appealing.