region test: fix memory access for host version
Using cuda::memory::host::vector results in an "invalid device context"
exception. (Apparently, this error did not occur in the 1.0.x branch
that used a previous version of cuda_wrapper.)

As a temporary solution, we copy to std::vector.

A much better way would be an overloaded function that handles the
various array types and copies them to host memory if needed.