Fix geometries bug
The geometry classes used to have a #ifdef __CUDACC__ around halmd::logger
inside them (introduced in 4c1350d03). This could lead to problems when
passing the class to a function compiled by nvcc (and not the host comiler)
due to the different sizes of the class.

This is also the bug for which a workaround was implemented for in 43b322bdd.
The bug had nothing to do with boost::function but somehow still got "fixed"
by using std::function.

This commit removes the logger from the class and the constructor. Logging of
the information can now instead be done by calling the ::log() function of
the geometry.