Change the configure_kernel() algorithm
configure_kernel() used to only try one calculated guess for the CUDA
dimensions and if this one didn't match it returned default_dim.

The new code starts with the minimum block size.
When the number of threads is fixed it tries different block sizes until
the maximum block size or the minimum grid size is exceeded.
If the number of threads is not fixed the minimum block size is used and
more threads may be configured than needed.