configure_kernel: implement new (soft) constraints
max_grid_size is close to infinity and thus useless (at least on a GTX 1070)

min_grid_size is not a constraint at all, instead it hints at the
physical dimensions of the current device. So we aim at grid sizes that
are multiples of min_grid_size or slightly smaller.

The case of a fixed thread number is a difficult algebraic problem. We
scan all possible block sizes and calculate a cost function for the
estimated runtime. Adding a threshold factor to the minimisation loop
helps favouring larger block sizes.