gpu/potentials/pair: Recreate textures parameters before launching kernels
When not recreating (or rebinding with the old API) the texture before
launching the kernel the texture of the parent potential (not the cutoff)
would start returning zero (when calling tex1Dfetch) after many
kernel launches.

It is unclear what causes this bug because when passing a pointer to
g_param_ directly and accessing it the data is fine. Also for some reason
the problem only arises in the parent potential and not in the cutoff
and the problem doesn't arise at all when not using a cutoff.

In addition only the first 8 bytes were returning zero when tested with
lennard_jones potential with a shifted cutoff.

The problem also doesn't arise when making a big enough GPU allocation
before starting the simulation (e.g. when iterating through the particle
positions on the host).