Add flag to neighbours to use one or 32 threads per particle in pair_trunc
Add the option 'unroll_force_loop' to be able to switch between one thread
per particle and 32 threads per particle in pair_trunc.

The flag is added to the neighbours because the matrices from the
neighbour lists need to be built different for pair_trunc to be able to
make use of coalesced access.