cmake: Ignore Thrust CUB version check
The Thrust version included in CUDA 11.0 starts checking whether the
CUB version being used is the one contained in CUDA. Because we are
using our own CUB version this check fails. This commit defines
THRUST_IGNORE_CUB_VERSION_CHECK to ignore this check when the
CUDA version is greater or equal to 11.0.