geometries.cylinder: fix potential issue with class size
Some parameters were stored as class attributes for logging only in the
C++ (host) definition of the class.

Having different class sizes in the C++ and CUDA definitions is a very
bad idea as it may lead to undefined behaviour if an instance of the
class is passed between host and GPU, e.g., when passed as an argument
in a CUDA kernel call. (This was the reason why the logger has been
factored out from the class.)

The extra logging parameters (original axis, unsquared radius) consume
4 floats = 32 bytes. Wrapping them in a struct, allocated on the heap, is
also not a good solution as passing shared_ptr or unique_ptr to CUDA has
its own issues. (And unique_ptr would delete the copy constructor.)
Managing this manually using new/delete is bad style and errorprone (the
geometry instance may be copied!)

So the simplest solution is to pass these extra parameters to the CUDA
kernel as well.

An alternative design would be moving the parameter logging completely
from the C++ part of the module to the Lua part, binding the parameter
values to a log() function attached to the Lua module.