observables.thermodynamics: fix registration of profiling timers
The timer 'mean_mass' does not exist, its (even two-fold) registration
lead to a segmentation fault when outputting the profiling results.

Revert to calling the local function profile(), introduced earlier. This
got lost in a recent merge commit.

In profile(), remove the safe-guard check about the existence of the
profiling timer. This is handled now in the class utility::profiler and
leads to an exception (which uncovers effectively dead code that would
otherwise be ignored silently.)

This fixes an issue introduced in commit 251d0b9f.