random/gpu/random_kernel.cu: Fix normal() function
The function should have used "v[k + GTDIM]" instead of "v[k + GTID]"
because this way some elements would not be written to and others
would be written to multiple times.

This commit now writes two consecutive elements in the array to avoid
writing out-of-bounds which could happen with the previous
implementation.