random/gpu: Add generic uniform() function
Add a new uniform() function which is generic over the RNG used.
This function differs from the old as it creates the floats by
using only 24 bits but in exchange creates uniform and unbiased
numbers. The method for createing these floats is described here:
https://lemire.me/blog/2017/02/28/how-many-floating-point-numbers-are-in-the-interval-01/

The uniform() functions from rand48 and MRG32k3a are removed.