observables.thermodynamics: pass volume as a module parameter
The volume of the particle group may be smaller than the simulation
domain, for example, if the particle group refers to a sub-domain
(geometric region). For such situations, the volume is provided as an
optional parameter to the thermodynamics module as a callable (nullary
function), which defaults to the box volume.

Implementationwise, the callable is stored as std::function (by copy).
The default is the empty function (nullptr), which is then replaced by a
call to box_->volume().