geometries.cylinder: a couple of serious bug fixes
The unit tests of the mdsim::particle_groups::region class failed
plainly for the cylinder geometry.

The predicate of the cylinder class was constructed wrongly: the
negation should use || instead of &&

In the test script, the cylinder length was not passed and not tested
for.

Remove class attributes enabled only for host compilation. This leads to
different class sizes for host and device compilation, which results in
plain wrongly passed kernel arguments. A similar issue was already fixed
in commit c06dad61f.

Rename class attribute 'offset' to 'centre'.