test/.../potentials/pair: obey cutoff for reference values
In the GPU tests, when computing reference values with the host
implementation, the cutoff was not respected. By chance (or by
construction), all particles were found to be within the cutoff
distance. However, the cutoff is measured in units of σ and was the same
for all species. When using different values of σ for different species,
this creates a problem that appeared first for the custom potential.

The issue is easily solved by checking within_range() of the potential.