observables::gpu::density_mode: fix issue with reduction kernels
The module was generating NaN values in the output. It seems that the
binary reduction kernel (complex_sum_) is at the root of the issue.
Reducing real and imaginary parts (cos, sin) separately solves the
problem. However, the issue requires further investigation. In
particular, there are is no unit test yet of the binary, ternary, ...
reduction kernels.