observables::gpu::density_mode_kernel: fix type of position array
For the position array, the declaration in the header file says 'float4
const*', but the kernel was defined with coalesced_vector_type. The
latter evaluates to float4 and float2 for dimension=3 and 2,
respectively. Strangely, the compiler accepted this. Yet it seems that
in two dimensions, every second particle was non-sense, maybe with
coordinates (0, species).