density_mode observables: use fixed_vector<..., 2> for complex numbers
The type std::complex is not enabled for CUDA (no __device__ keyword).
With halmd::fixed_vector we have addition, which is sufficient for our
purposes.

GPU arrays are of type float2 instead of fixed_vector<float, 2>. This
follows the overall pattern applied in HALMD, which has its origin in
the early times of CUDA. Presumingly, arrays to compound types (struct)
are permissable nowadays, so the conversion may not be necessary
anymore.