obserables::gpu::density_mode: inherit CUDA execution dimensions from particle
The compute kernel runs over all particles in the first place. So it
seems reasonable to benefit from potential optimisations of
mdsim::gpu::particle on the execution dimensions.

Caveat: the number of threads per block must be a power of 2 for
reduce_warp to work smoothly. Currently, only such values are returned
from mdsim::gpu::particle.