mdsim.potentials: fix size check of planar_wall parameter matrices
The potential method size() is queried by the force modules and is
compared with the max. number of particle species. The offset vector,
however, yields the number of walls. Now, we return the second dimension
of the epsilon array.

The bug triggers the following error message of the force module if
there are less walls than particle species':

[ERROR] size of potential coefficients less than number of particle species

The issue was not detected by the unit test, which uses 3 walls and
2 species.

Unit test: rename 'ntype' to 'nspecies'.