doc: switch to LuaLaTeX in Sphinx configuration
With recent Sphinx/LaTeX installations (e.g., on Debian 12), PDFLaTeX
results in a series of unknown UTF8 characters. Switching to LuaLaTeX
(almost) solves the problem.

Remove the global declaration of unusual UTF8 characters from the
configuration, since it is not needed with LuaLaTeX.

The Ångstrøm symbol Å is not known in math mode, so put it inside of
\text{} in doc/units.rst.in:

Missing character: There is no Å (U+00C5) in font cmmi10!