observables.dynamics: add self-intermediate scattering function (host backend)
The module combines similar code from the intermediate scattering
function (for the result type and the wavevector dependency), the
mean-square displacement (for the sample type and the loop over
particles), and the density modes (for the actual calculation and the
loop over wave vectors).

So far, only the host backend has been implemented. For simulations on
GPUs, the phase space sample is taken in host memory ("acquire"
function). Hence, this correlation function can also be computed on
GPUs, with slow performance.