Test fixed_vector_cuda_vector_converter: make use of dsfloat_cuda_vector
The test should be as close as possible to the actual use case in the
code. In addition, dsfloat_cuda_vector simplifies the assignment to
float4 tuples in case of double-single fixed_vector.

Extensive debugging suggests that the test failure is due to a compiler
bug of GCC 12. The lower significant part of the dsfloat numbers is not
written to get<1>(h_input[i]), which instead is filled by zero.

In test/unit/numeric/blas/fixed_vector_cuda_vector_converter.cpp:425:

        h_input[i] <<= tie(u, v);

Strangely, the identical lines in the 'basic' test before work.

https://my.cdash.org/testSummary.php?project=554&name=unit/numeric/fixed_vector/cuda_vector_converter&date=2025-04-03