dynamics::blocking_scheme: make computation of sampling intervals more robust
The design decision to pass resolution and maximum lag time in units of
time and not steps makes it necessary to convert these quantities to
integration steps, via clock::timestep(). However, this may incur subtle
roundoff errors that would lead to a finest sampling interval of s = 0.

The issue happened when passing timestep = 0.015 and resolution = 0.015.
A possible reason may be compiler optimisations and 80-bit vs. 64-bit
floating-point registers.