integrators::verlet: lock forces during integrate() step
If multiple integrators are used, on the same or different instances of
mdsim::particle, forces must not be updated despite a change of
positions. With the new locking mechanism for particle arrays, we ensure
that the force values during integrate() correspond to the positions
before the first half step.

Add slot functions prepend_integrate() and append_integrate(), which are
connected in verlet.lua to the corresponding signals of mdsim::core.